Suggestion: to win new translators
Advantages: The language files stay outside of the plugin.
There does not have to be a new version of the plugin to update the language files.
Disadvantages: The translations for each language will be unlocked by the Polyglot team, unless there is a PTE (Project Translate Editor).
PTE rights can be applied separately for each language by the plugin developer.
Quite simply, if one translates to 95%, WP will generate a language file that can be downloaded. It is stored outside the plugin. If this file exists, the sparing file in the plugin will no longer be loaded.